Add files via upload
Users can enter search queries to discover images, and a "Load More" button enables the retrieval of additional images from the Pexels API. The gallery displays images in a responsive grid layout, and clicking on an image opens a detailed lightbox with the photographer's name and a download option. The code efficiently integrates API calls, and event listeners enhance user interactions, offering a seamless and engaging experience for exploring and downloading images. The implementation also includes pagination for convenient navigation through the gallery. Overall it  provides a visually appealing and functional image gallery with modern features.